Fix Error on Delete Category & Sub Category

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
 
@@ -96,6 +97,15 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findOrFail($id);
+        $subCategories = SubCategory::where('category_id', $category->id)->count();
+
+        if ($subCategories > 0) {
+            return response([
+                'status' => 'error',
+                'message' => 'Category contains sub item, please delete sub item first!'
+            ]);
+        }
+
         $category->delete();
 
         return response([
